Add files via upload
Python Video to Audio Converter Project:-
Video to audio converter is an application that allows you to convert your video to audio and video to mp3. In this application, you have to select a file and when you are done with selecting your video file just press the convert button. This video to audio converter application will then start converting video to audio and your audio file will be available in the same folder. Let’s develop Video to Audio Converter project using Python.